Refactored test file names to match pytest requirements
